Fix some of the old references to getCanonicalSequence. These were problems left over from when we merged in forge_master back to the main branch.

+++ src/main/java/contrail/avro/BuildGraphAvro.java
@@ -562,7 +562,7 @@ public class BuildGraphAvro extends Stage
       JobClient.runJob(conf);
       long endtime = System.currentTimeMillis();
 
-      float diff = (float) (((float) (endtime - starttime)) / 1000.0);
+      float diff = (float) ((endtime - starttime) / 1000.0);
 
       sLogger.info("Runtime: " + diff + " s");
     }
